Wait for tests to complete before returning from run_tests().
Previously we waited for the tests to complete before accessing the results,
but now that the code for running all the tests is moved to the Tester class,
we can do it this way instead.

This makes the code clearer, and will make it easier to make the code
interruptible.